backend/pyproject.toml
[tool.poetry]
name = "nft_market"
version = "0.1.0"
description = ""
authors = ["Sebastian Gula <sebastian.gula@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1.7"
djangorestframework = "3.12.2"
pyteal = {git = "https://github.com/algorand/pyteal.git", rev = "77b2e4452e443d449474d4dee1776fab3eb66f66"}
PyYAML = "^5.4.1"
py-algorand-sdk = "^1.4.1"
django-cors-headers = "^3.7.0"
Pillow = "^8.1.2"
django-filter = "^2.4.0"
celery = "^5.0.5"
uWSGI = "^2.0.19"
dj-database-url = "^0.5.0"
django-storages = {extras = ["google"], version = "^1.11.1"}
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
